# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_vision_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_vision_upload_post",
        "type": "object"
      },
...
      "CameraBody": {
        "description": "One frame from a webcam.",
        "properties": {
          "index": {
            "default": 0,
            "title": "Index",
            "type": "integer"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          }
        },
        "title": "CameraBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoBody": {
        "description": "Sample frames from a video file.",
        "properties": {
          "frames": {
            "default": 3,
            "title": "Frames",
            "type": "integer"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          }
        },
        "title": "VideoBody",
        "type": "object"
      },
      "VisionAnalyseBody": {
        "description": "An image to look at: either a path on disk or an inline data URL.",
        "properties": {
          "data_url": {
            "default": "",
            "title": "Data Url",
            "type": "string"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          }
        },
        "title": "VisionAnalyseBody",
        "type": "object"
      },
      "VisionOcrBody": {
        "description": "Text extraction from an image or a document.",
        "properties": {
          "data_url": {
            "default": "",
            "title": "Data Url",
            "type": "string"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          }
        },
        "title": "VisionOcrBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first personal agent: chat, missions, memory, tools, governance.",
    "title": "Natasha API",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/vision/analyse": {
      "post": {
        "description": "Analyse an image already on disk (inside the allowed roots).",
        "operationId": "analyse_api_vision_analyse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionAnalyseBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analyse Api Vision Analyse Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyse",
        "tags": [
          "vision"
        ]
      }
    },
    "/api/vision/camera": {
      "post": {
        "description": "Grab one frame from a webcam and analyse it. Requires OpenCV and a camera.",
        "operationId": "camera_api_vision_camera_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Camera Api Vision Camera Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Camera",
        "tags": [
          "vision"
        ]
      }
    },
...
    "/api/vision/documents": {
      "post": {
        "description": "Read a document (pdf/docx/xlsx/pptx/csv/image/zip/text) as untrusted content.",
        "operationId": "documents_api_vision_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionOcrBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Documents Api Vision Documents Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents",
        "tags": [
          "vision"
        ]
      }
    },
    "/api/vision/ocr": {
      "post": {
        "description": "Extract text from an image, screenshot or document without a vision model.",
        "operationId": "ocr_api_vision_ocr_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionOcrBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ocr Api Vision Ocr Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ocr",
        "tags": [
          "vision"
        ]
      }
    },
    "/api/vision/screen": {
      "post": {
        "description": "Screenshot the desktop (through the computer-use controller, so policy still applies) and look at it.",
        "operationId": "screen_api_vision_screen_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionAnalyseBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Screen Api Vision Screen Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Screen",
        "tags": [
          "vision"
        ]
      }
    },
    "/api/vision/ui": {
      "post": {
        "description": "Interpret a UI: combine the screenshot with the accessibility tree into one untrusted block.",
        "operationId": "ui_understanding_api_vision_ui_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionAnalyseBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ui Understanding Api Vision Ui Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Understanding",
        "tags": [
          "vision"
        ]
      }
    },
    "/api/vision/upload": {
      "post": {
        "description": "Upload an image and analyse it in one step (what the UI uses).",
        "operationId": "upload_api_vision_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "prompt",
            "required": false,
            "schema": {
              "default": "",
              "title": "Prompt",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_vision_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Api Vision Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "vision"
        ]
      }
    },
    "/api/vision/video": {
      "post": {
        "description": "Sample frames from a video with ffmpeg and analyse each one.",
        "operationId": "video_api_vision_video_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Video Api Vision Video Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Video",
        "tags": [
          "vision"
        ]
      }
    },
...
  }
}